refactor: change the structure of the batabase access layer
