+++ internal/repository/repository.go
@@ -1,0 +1,7 @@
+package repository
+
+import "github.com/CleysonPH/reading-tracker/internal/model"
+
+type BookRepository interface {
+	All(q string) ([]*model.Book, error)
+}
